+++ nakadi.go
@@ -22,7 +22,6 @@ import (
 	"time"
 
 	"github.com/cenkalti/backoff/v3"
-	"github.com/opentracing/opentracing-go"
 	"github.com/pkg/errors"
 )
 
@@ -43,17 +42,15 @@ type Client struct {
 	httpStreamClient *http.Client
 }
 
+// Middleware provides a chainable http.RoundTripper middleware that can be used
+// to hook into requests e.g. for logging or tracing purposes.
+type Middleware func(transport *http.Transport) http.RoundTripper
+
 // ClientOptions contains all non mandatory parameters used to instantiate the Nakadi client.
 type ClientOptions struct {
 	TokenProvider     func() (string, error)
 	ConnectionTimeout time.Duration
-	TracingOptions    TracingOptions
-}
-
-type TracingOptions struct {
-	Tracer        opentracing.Tracer
-	SpanName      string
-	ComponentName string
+	Middleware        Middleware
 }
 
 func (o *ClientOptions) withDefaults() *ClientOptions {
@@ -63,6 +60,9 @@ func (o *ClientOptions) withDefaults() *ClientOptions {
 	}
 	if copyOptions.ConnectionTimeout == 0 {
 		copyOptions.ConnectionTimeout = defaultTimeOut
+	}
+	if copyOptions.Middleware == nil {
+		copyOptions.Middleware = func(transport *http.Transport) http.RoundTripper { return transport }
 	}
 	return &copyOptions
 }
@@ -77,8 +77,8 @@ func New(url string, options *ClientOptions) *Client {
 		nakadiURL:        url,
 		timeout:          options.ConnectionTimeout,
 		tokenProvider:    options.TokenProvider,
-		httpClient:       newHTTPClient(options.ConnectionTimeout, &options.TracingOptions),
-		httpStreamClient: newHTTPStream(options.ConnectionTimeout, &options.TracingOptions)}
+		httpClient:       newHTTPClient(options.ConnectionTimeout, options.Middleware),
+		httpStreamClient: newHTTPStream(options.ConnectionTimeout)}
 
 	return client
 }
